# PERF: ArrowExtensionArray._from_sequence_of_strings(dtype=pa.bool_())