- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [x] Added an entry in the latest `doc/source/whatsnew/v2.1.0.rst` file if fixing a bug or adding a new feature.

`ArrowExtensionArray._from_sequence_of_strings` currently depends on `BooleanArray._from_sequence_of_strings` to parse strings for `boolean[pyarrow]`. This removes the dependency on `BooleanArray` and uses pyarrow casting to accomplish the same thing. I've adding some logic around pyarrow's casting to match existing behavior of `BooleanArray` (assuming we want to maintain existing behavior?)

```
import pyarrow as pa
import numpy as np
from pandas.core.arrays.arrow.array import ArrowExtensionArray

N = 1_000_000
strings_np = np.random.choice(["true", "false", None], N, replace=True)
strings_pa = pa.array(strings_np, type=pa.string())

%timeit ArrowExtensionArray._from_sequence_of_strings(strings_np, dtype=pa.bool_())

# 200 ms ± 4.72 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)    -> main
# 54.5 ms ± 842 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)  -> PR

%timeit ArrowExtensionArray._from_sequence_of_strings(strings_pa, dtype=pa.bool_())

# 298 ms ± 3.45 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)    -> main
# 30.8 ms ± 1.1 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)  -> PR
```